requirements.txt
audio2numpy==0.1.2
pescador==2.1.0
filelock==3.0.12
librosa==0.8.0
tabulate==0.8.9
sortedcontainers==2.2.2
spacy==2.3.0
pandas==1.1.3
sympy==1.6.2
pydub==0.25.1
torch==1.7.1
docopt==0.6.2
simplejson==3.17.2
typing_extensions==3.7.4.3
tqdm==4.50.2
torchaudio==0.7.0a0+a853dff
SoundFile==0.10.3.post1
matplotlib==3.3.2
optuna==2.5.0
scipy==1.5.2
dataclasses==0.8
ipython==7.24.1
pyroomacoustics==0.4.3
PyYAML==5.4.1
scikit_learn==0.24.2
typer==0.3.2
xarray==0.18.2
